Move default_content_filter to signals.py for consistency
